Moved `keyBuffer` to `BrsDevice` and returns next key using `updateKeysBuffer`
